requirements.txt
SpeechRecognition
Pillow
fpdf
PyAudio